Templater Scripts - Folder Note Item List Automation
This script will create a template to be able to generate an item list, with folder note links, upon creation of a folder note.
Instead of having to create the item list manually, copying it over, creating the links to those folders, this script will do most of the work for you by generating an item list with links to the children folders. This means whenever you create a new child folder, you can delete the parent's folder note and create a new one and the link the new folder note will be added to the folder note.
The pictures of the children folders will have a placeholder: "folder-note-child-name.jpg". The only thing you need to do is add pictures with the folder names to the library asset folder. If not there will be no picture, but links will still work.